refactor: reorganize common parameters and scripts into utility_scripts module
