+++ starter/src/modules_scripts.ts
@@ -3,24 +3,7 @@ import {
   firebaseIOSParameters,
   firebaseWebParameters,
 } from './common-params';
-import { Parameter, Script } from './interfaces';
-
-// Common parameters available across scripts and modules
-export const commonParameters: Parameter[] = [
-  {
-    key: 'platform',
-    question: 'Which platform(s) are you targeting?',
-    type: 'select',
-    choices: ['ios', 'android', 'web'],
-    platform: ['android', 'ios', 'web'],
-  },
-  {
-    key: 'ensemble_version',
-    question: 'Which version of ensemble are you using?',
-    type: 'text',
-    platform: ['android', 'ios', 'web'],
-  },
-];
+import { Script } from './interfaces';
 
 // Modules (called with `enable` command)
 export const modules: Script[] = [
@@ -294,34 +277,6 @@ export const modules: Script[] = [
         key: 'webGoogleMapsApiKey',
         question: 'Please provide your Google Maps API key for Web ',
         platform: ['web'],
-        type: 'text',
-      },
-    ],
-  },
-];
-
-// Custom Scripts (standalone Dart scripts)
-export const scripts: Script[] = [
-  {
-    name: 'generateKeystore',
-    path: 'scripts/generate_keystore.dart',
-    parameters: [
-      {
-        key: 'storePassword',
-        question: 'Please provide the store password: ',
-        platform: ['android'],
-        type: 'text',
-      },
-      {
-        key: 'keyPassword',
-        question: 'Please provide the key password: ',
-        platform: ['android'],
-        type: 'text',
-      },
-      {
-        key: 'keyAlias',
-        question: 'Please provide the key alias: ',
-        platform: ['android'],
         type: 'text',
       },
     ],
